Rename sketch file to match folder name
The Arduino IDE requires the sketch folder name to match the filename of the primary sketch file. This includes case matching on a filename case-sensitive OS like Linux. This change causes the example sketch to be accessible via the Arduino IDE's File > Examples > LIBRARYNAME menu after the library is installed.